+++ src/main/java/io/contek/invoker/okx/api/common/_LoanDetails.java
@@ -1,7 +1,6 @@
 package io.contek.invoker.okx.api.common;
 
 import javax.annotation.concurrent.NotThreadSafe;
-import java.math.BigDecimal;
 
 @NotThreadSafe
 public class _LoanDetails {
@@ -9,7 +8,7 @@ public class _LoanDetails {
   public String instId;
   public String mgnMode;
   public String mgnCcy;
-  public BigDecimal maxLoan;
+  public Double maxLoan;
   public String ccy;
   public String side;
 }
