+++ src/main/java/io/contek/invoker/okx/api/common/_Trade.java
@@ -1,15 +1,14 @@
 package io.contek.invoker.okx.api.common;
 
 import javax.annotation.concurrent.NotThreadSafe;
-import java.math.BigDecimal;
 
 @NotThreadSafe
 public class _Trade {
 
   public String instId;
   public String tradeId;
-  public BigDecimal px;
-  public BigDecimal sz;
+  public Double px;
+  public Double sz;
   public String side;
   public long ts;
   public long traceNano = System.nanoTime();
